fix: some fixes with the database types due to overflowing numbers

+++ Api/AntScape/src/main/java/es/iespto/algyjmcg/AntScape/infrastructure/adapter/primary/v2/NestLevelV2Controller.java
@@ -35,7 +35,7 @@ public class NestLevelV2Controller {
 
 		findById.setProduction(res.doubleValue());
 
-		findById.setCost(findById.getCost()*findById.getMultiplier().doubleValue());
+		findById.setCost(findById.getCost()*findById.getMultiplier().floatValue());
 		
 		boolean update = mainService.update(findById);
 		
@@ -88,7 +88,7 @@ public class NestLevelV2Controller {
 
 class NestLevelSaveInputDTO {
 	private Integer id;
-	private Double cost;
+	private Float cost;
 	private Integer level;
 	private BigDecimal multiplier;
 	private String name;
@@ -105,10 +105,10 @@ class NestLevelSaveInputDTO {
 	public void setId(Integer id) {
 		this.id = id;
 	}
-	public Double getCost() {
+	public Float getCost() {
 		return cost;
 	}
-	public void setCost(Double cost) {
+	public void setCost(Float cost) {
 		this.cost = cost;
 	}
 	public Integer getLevel() {
